refactor(web): move useRefreshReviews under github domain and widen dropdown
Relocates the hook to hooks/domains/github/ to match the project's
domain-organization convention for GitHub-related hooks, and delegates
the watches fetch to the existing useReviewWatches hook so we inherit
its store-level loading guard rather than rolling our own module-scoped
in-flight flag. Also switches the overflow DropdownMenuContent from
the default trigger-width constraint to w-fit so items like "Check for
PRs to review" render on one line instead of wrapping to three.
